+++ app/static/js/show-spinners.js
@@ -1,0 +1,5 @@
+// Make spinners appear.
+function showSpinners() {
+  const spinners = document.getElementById('spinners');
+  spinners.style.display = 'block';
+}